--- a/openssl/goopenssl.h
+++ b/openssl/goopenssl.h
@@ -35,6 +35,9 @@ void go_openssl_load_functions(void* handle, const void* v1_0_sentinel, const vo
 #define GO_AES_ENCRYPT 1
 #define GO_AES_DECRYPT 0
 
+typedef void* GO_EVP_CIPHER_PTR;
+typedef void* GO_EVP_CIPHER_CTX_PTR;
+
 // Define pointers to all the used OpenSSL functions.
 // Calling C function pointers from Go is currently not supported.
 // It is possible to circumvent this by using a C function wrapper.
@@ -68,21 +71,21 @@ FOR_ALL_OPENSSL_FUNCTIONS
 // These wrappers allocate out_len on the C stack, and check that it matches the expected
 // value, to avoid having to pass a pointer from Go, which would escape to the heap.
 static inline int
-go_openssl_EVP_EncryptUpdate_wrapper(EVP_CIPHER_CTX *ctx, uint8_t *out, const uint8_t *in, int in_len)
+go_openssl_EVP_EncryptUpdate_wrapper(GO_EVP_CIPHER_CTX_PTR ctx, uint8_t *out, const uint8_t *in, int in_len)
 {
     int len;
     return go_openssl_EVP_EncryptUpdate(ctx, out, &len, in, in_len);
 }
 
 static inline int
-go_openssl_EVP_DecryptUpdate_wrapper(EVP_CIPHER_CTX *ctx, uint8_t *out, const uint8_t *in, int in_len)
+go_openssl_EVP_DecryptUpdate_wrapper(GO_EVP_CIPHER_CTX_PTR ctx, uint8_t *out, const uint8_t *in, int in_len)
 {
     int len;
     return go_openssl_EVP_DecryptUpdate(ctx, out, &len, in, in_len);
 }
 
 static inline int
-go_openssl_EVP_CipherUpdate_wrapper(EVP_CIPHER_CTX *ctx, uint8_t *out, const uint8_t *in, int in_len)
+go_openssl_EVP_CipherUpdate_wrapper(GO_EVP_CIPHER_CTX_PTR ctx, uint8_t *out, const uint8_t *in, int in_len)
 {
     int len;
     return go_openssl_EVP_CipherUpdate(ctx, out, &len, in, in_len);