--- a/openssl/openssl_funcs.h
+++ b/openssl/openssl_funcs.h
@@ -87,13 +87,13 @@ DEFINEFUNC(size_t, HMAC_CTX_copy, (HMAC_CTX *dest, HMAC_CTX *src), (dest, src)) 
 DEFINEFUNC_1_1(void, HMAC_CTX_free, (HMAC_CTX * arg0), (arg0)) \
 DEFINEFUNC_1_1(HMAC_CTX*, HMAC_CTX_new, (void), ()) \
 DEFINEFUNC_1_1(void, HMAC_CTX_reset, (HMAC_CTX * arg0), (arg0)) \
-DEFINEFUNC(EVP_CIPHER_CTX *, EVP_CIPHER_CTX_new, (void), ()) \
-DEFINEFUNC(int, EVP_CIPHER_CTX_set_padding, (EVP_CIPHER_CTX *x, int padding), (x, padding)) \
+DEFINEFUNC(GO_EVP_CIPHER_CTX_PTR, EVP_CIPHER_CTX_new, (void), ()) \
+DEFINEFUNC(int, EVP_CIPHER_CTX_set_padding, (GO_EVP_CIPHER_CTX_PTR x, int padding), (x, padding)) \
 DEFINEFUNC(int, EVP_CipherInit_ex, \
-           (EVP_CIPHER_CTX * ctx, const EVP_CIPHER *type, ENGINE *impl, const unsigned char *key, const unsigned char *iv, int enc), \
+           (GO_EVP_CIPHER_CTX_PTR ctx, const GO_EVP_CIPHER_PTR type, ENGINE *impl, const unsigned char *key, const unsigned char *iv, int enc), \
            (ctx, type, impl, key, iv, enc)) \
 DEFINEFUNC(int, EVP_CipherUpdate, \
-           (EVP_CIPHER_CTX * ctx, unsigned char *out, int *outl, const unsigned char *in, int inl), \
+           (GO_EVP_CIPHER_CTX_PTR ctx, unsigned char *out, int *outl, const unsigned char *in, int inl), \
            (ctx, out, outl, in, inl)) \
 DEFINEFUNC(BIGNUM *, BN_new, (void), ()) \
 DEFINEFUNC(void, BN_free, (BIGNUM * arg0), (arg0)) \
@@ -129,31 +129,31 @@ DEFINEFUNC_1_1(void, RSA_get0_key, \
     (const RSA *rsa, const BIGNUM **n, const BIGNUM **e, const BIGNUM **d), \
     (rsa, n, e, d)) \
 DEFINEFUNC(int, EVP_EncryptInit_ex, \
-    (EVP_CIPHER_CTX *ctx, const EVP_CIPHER *type, ENGINE *impl, const unsigned char *key, const unsigned char *iv), \
+    (GO_EVP_CIPHER_CTX_PTR ctx, const GO_EVP_CIPHER_PTR type, ENGINE *impl, const unsigned char *key, const unsigned char *iv), \
     (ctx, type, impl, key, iv)) \
 DEFINEFUNC(int, EVP_EncryptUpdate, \
-    (EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl, const unsigned char *in, int inl), \
+    (GO_EVP_CIPHER_CTX_PTR ctx, unsigned char *out, int *outl, const unsigned char *in, int inl), \
     (ctx, out, outl, in, inl)) \
 DEFINEFUNC(int, EVP_EncryptFinal_ex, \
-    (EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl), \
+    (GO_EVP_CIPHER_CTX_PTR ctx, unsigned char *out, int *outl), \
     (ctx, out, outl)) \
 DEFINEFUNC(int, EVP_DecryptUpdate, \
-    (EVP_CIPHER_CTX *ctx, unsigned char *out, int *outl, const unsigned char *in, int inl),	(ctx, out, outl, in, inl)) \
-DEFINEFUNC(int, EVP_DecryptFinal_ex, (EVP_CIPHER_CTX *ctx, unsigned char *outm, int *outl),	(ctx, outm, outl)) \
-DEFINEFUNC(const EVP_CIPHER*, EVP_aes_128_gcm, (void), ()) \
-DEFINEFUNC(const EVP_CIPHER*, EVP_aes_128_cbc, (void), ()) \
-DEFINEFUNC(const EVP_CIPHER*, EVP_aes_128_ctr, (void), ()) \
-DEFINEFUNC(const EVP_CIPHER*, EVP_aes_128_ecb, (void), ()) \
-DEFINEFUNC(const EVP_CIPHER*, EVP_aes_192_gcm, (void), ()) \
-DEFINEFUNC(const EVP_CIPHER*, EVP_aes_192_cbc, (void), ()) \
-DEFINEFUNC(const EVP_CIPHER*, EVP_aes_192_ctr, (void), ()) \
-DEFINEFUNC(const EVP_CIPHER*, EVP_aes_192_ecb, (void), ()) \
-DEFINEFUNC(const EVP_CIPHER*, EVP_aes_256_cbc, (void), ()) \
-DEFINEFUNC(const EVP_CIPHER*, EVP_aes_256_ctr, (void), ()) \
-DEFINEFUNC(const EVP_CIPHER*, EVP_aes_256_ecb, (void), ()) \
-DEFINEFUNC(const EVP_CIPHER*, EVP_aes_256_gcm, (void), ()) \
-DEFINEFUNC(void, EVP_CIPHER_CTX_free, (EVP_CIPHER_CTX* arg0), (arg0)) \
-DEFINEFUNC(int, EVP_CIPHER_CTX_ctrl, (EVP_CIPHER_CTX *ctx, int type, int arg, void *ptr), (ctx, type, arg, ptr)) \
+    (GO_EVP_CIPHER_CTX_PTR ctx, unsigned char *out, int *outl, const unsigned char *in, int inl),	(ctx, out, outl, in, inl)) \
+DEFINEFUNC(int, EVP_DecryptFinal_ex, (GO_EVP_CIPHER_CTX_PTR ctx, unsigned char *outm, int *outl),	(ctx, outm, outl)) \
+DEFINEFUNC(const GO_EVP_CIPHER_PTR, EVP_aes_128_gcm, (void), ()) \
+DEFINEFUNC(const GO_EVP_CIPHER_PTR, EVP_aes_128_cbc, (void), ()) \
+DEFINEFUNC(const GO_EVP_CIPHER_PTR, EVP_aes_128_ctr, (void), ()) \
+DEFINEFUNC(const GO_EVP_CIPHER_PTR, EVP_aes_128_ecb, (void), ()) \
+DEFINEFUNC(const GO_EVP_CIPHER_PTR, EVP_aes_192_gcm, (void), ()) \
+DEFINEFUNC(const GO_EVP_CIPHER_PTR, EVP_aes_192_cbc, (void), ()) \
+DEFINEFUNC(const GO_EVP_CIPHER_PTR, EVP_aes_192_ctr, (void), ()) \
+DEFINEFUNC(const GO_EVP_CIPHER_PTR, EVP_aes_192_ecb, (void), ()) \
+DEFINEFUNC(const GO_EVP_CIPHER_PTR, EVP_aes_256_cbc, (void), ()) \
+DEFINEFUNC(const GO_EVP_CIPHER_PTR, EVP_aes_256_ctr, (void), ()) \
+DEFINEFUNC(const GO_EVP_CIPHER_PTR, EVP_aes_256_ecb, (void), ()) \
+DEFINEFUNC(const GO_EVP_CIPHER_PTR, EVP_aes_256_gcm, (void), ()) \
+DEFINEFUNC(void, EVP_CIPHER_CTX_free, (GO_EVP_CIPHER_CTX_PTR arg0), (arg0)) \
+DEFINEFUNC(int, EVP_CIPHER_CTX_ctrl, (GO_EVP_CIPHER_CTX_PTR ctx, int type, int arg, void *ptr), (ctx, type, arg, ptr)) \
 DEFINEFUNC(EVP_PKEY *, EVP_PKEY_new, (void), ()) \
 DEFINEFUNC_RENAMED(int, EVP_PKEY_get_size, EVP_PKEY_size, (const EVP_PKEY *pkey), (pkey)) \
 DEFINEFUNC(void, EVP_PKEY_free, (EVP_PKEY * arg0), (arg0)) \