This PR contains two complementary AES performance improvements:

## Using AES wrappers

Commit: 99e30441c570331aa6940ae37ab02681c87a2b73

Several AES functions and methods were calling EVP_EncryptUpdate, EVP_DecryptUpdate and EVP_CipherUpdate without using the third parameter, `out_len`, which is an integer pointer filled by OpenSSL with the length of the output buffer.

`out_len` escapes to the heap, which puts unnecessary garbage collector pressure and degrades performance. To avoid that we are now allocating `out_len` in the C stack and just discarding the result.

## Declare pointer types in C

Commit: 9356effed82c0633877fdf1c214026c68f0f3454

We are storing EVP_CIPHER_CTX and EVP_CIPHER objects as Go pointers even though these are opaque C pointers which Go doesn't need to know they are pointers as it is not creating them nor doing any special operation, everything happens in the C scope. If we store those objects as `void*` typedefs we will have more chances that Go escape analysis will keep them on the stack.

A side-effect of this change is that we rely less on types defined in OpenSSL headers, which goes in the right direction. The "north star" is to eliminate the prerequisite of needing OpenSSL development headers in order to build this package.

As a side note, boringcrypto does not suffer so much from these allocations because BoringSSL does not use opaque pointers as function arguments but ordinary structs, which are are easier to allocate in the stack.

# Results

AES encryption and decryption no longer allocates and has 70% more throughput.
AES seal and open allocates 75% less and has ~20% more throughput.

```txt
name                  old time/op    new time/op     delta
AESEncrypt-4             171ns ± 3%      102ns ± 6%   -40.22%  (p=0.000 n=10+10)
AESDecrypt-4             173ns ± 4%      101ns ± 6%   -41.36%  (p=0.000 n=10+10)
AESGCM/Open-128-64-4    1.28µs ± 2%     1.07µs ± 2%   -16.65%  (p=0.000 n=10+9)
AESGCM/Seal-128-64-4    1.31µs ± 3%     1.07µs ± 4%   -18.69%  (p=0.000 n=10+10)

name                  old speed      new speed       delta
AESEncrypt-4          93.5MB/s ± 3%  156.5MB/s ± 6%   +67.45%  (p=0.000 n=10+10)
AESDecrypt-4          92.7MB/s ± 4%  158.1MB/s ± 5%   +70.61%  (p=0.000 n=10+10)
AESGCM/Open-128-64-4  50.0MB/s ± 2%   59.7MB/s ± 5%   +19.31%  (p=0.000 n=10+10)
AESGCM/Seal-128-64-4  48.7MB/s ± 3%   59.9MB/s ± 4%   +22.93%  (p=0.000 n=10+10)

name                  old B/op   new B/op    delta
AESEncrypt-4             16.0B ± 0%       0.0B       -100.00%  (p=0.000 n=10+10)
AESDecrypt-4             16.0B ± 0%       0.0B       -100.00%  (p=0.000 n=10+10)
AESGCM/Open-128-64-4      120B ± 0%        32B ± 0%   -73.33%  (p=0.000 n=10+10)
AESGCM/Seal-128-64-4      112B ± 0%        32B ± 0%   -71.43%  (p=0.000 n=10+10)

name                  old allocs/op  new allocs/op   delta
AESEncrypt-4              2.00 ± 0%       0.00       -100.00%  (p=0.000 n=10+10)
AESDecrypt-4              2.00 ± 0%       0.00       -100.00%  (p=0.000 n=10+10)
AESGCM/Open-128-64-4      12.0 ± 0%        3.0 ± 0%   -75.00%  (p=0.000 n=10+10)
AESGCM/Seal-128-64-4      11.0 ± 0%        3.0 ± 0%   -72.73%  (p=0.000 n=10+10)
```