@qmuntal 2022-02-22
AES encrypt and decrypt performance are almost maxed out, even though we are 10 times slower than standard crypto: 10ns/op vs 100ns/op in my dev environment.

If we generate a cpu profile from BenchmarkAESEncrypt we can see that the cgo call and cgo boilerplate accounts for the 90% of the time, and we can't reduce the cgo calls more.

```bash
$ go test -run=- -bench=BenchmarkAESEncrypt -memprofile mem.out -cpuprofile cpu.out
$ go tool pprof cpu.out 
File: openssl.test
Build ID: 998359c183439631807bc0a0210fe2d55a6d6072
Type: cpu
Time: Feb 22, 2022 at 10:34am (UTC)
Duration: 1.40s, Total samples = 1.28s (91.18%)
Entering interactive mode (type "help" for commands, "o" for options)
(pprof) top
Showing nodes accounting for 1.27s, 99.22% of 1.28s total
Showing top 10 nodes out of 15
      flat  flat%   sum%        cum   cum%
     0.80s 62.50% 62.50%      1.23s 96.09%  runtime.cgocall
     0.15s 11.72% 74.22%      0.22s 17.19%  runtime.exitsyscallfast
     0.09s  7.03% 81.25%      0.11s  8.59%  runtime.casgstatus
     0.09s  7.03% 88.28%      0.42s 32.81%  runtime.exitsyscall
     0.06s  4.69% 92.97%      0.06s  4.69%  runtime.wirep
     0.02s  1.56% 94.53%      1.27s 99.22%  github.com/microsoft/go-crypto-openssl/openssl.(*aesCipher).Encrypt
     0.02s  1.56% 96.09%      1.25s 97.66%  github.com/microsoft/go-crypto-openssl/openssl._Cfunc_go_openssl_EVP_EncryptUpdate_wrapper
     0.02s  1.56% 97.66%      0.02s  1.56%  runtime.nanotime (inline)
     0.01s  0.78% 98.44%      1.28s   100%  github.com/microsoft/go-crypto-openssl/openssl.BenchmarkAESEncrypt
     0.01s  0.78% 99.22%      0.01s  0.78%  runtime.exitsyscallfast_reacquired
```