@qmuntal 2022-02-22
AES Open and Seal still have room for improvement if we batch 5 cgo calls into one, but it would mean moving logic from Go to C, which I would try to avoid for safety sake.