@qmuntal 2022-02-22
We run ~7% faster on go1.18rc1 compared to go1.17.7, which is strange because it is a significant speed burst and go1.18rc1 haven't done any major improvement in linux-amd64 ABI or cgo boilerplate, at least that I know.